feat: added utils
