+++ src/utils/updateTextStyles.js
@@ -1,0 +1,8 @@
+export const updateTextStyles = (bold, italic) => {
+  return {
+    padding: 10,
+    fontSize: 16,
+    fontWeight: bold ? 'bold' : 'normal',
+    fontStyle: italic ? 'italic' : 'normal',
+  };
+};
